fix(ui): 批T 审查修复（PR #2608）——image 线宽/关闭钮坐标/图标悬浮/标签框心
审查 1 BLOCKER + 2 MAJOR + 2 MINOR 全修：
- BLOCKER image 线宽：ItemInfo.image 是 u16，服务端漏强转只写 2 字节而
  客户端按 i32 读 4 字节——真实链路整包错位（mock 4 字节写所以 e2e 漏网）；
  改 (image as i32).to_le_bytes()
- MAJOR 关闭钮坐标：509 是 HelpDialog 宽面板的（536+509=1045 超出 1024
  视口，屏幕外点不到）；InspectDialog 是 (241,3)（MainDialogs.cs:2213）
- MAJOR 图标悬浮：icon_system 补 is_open 门控（关闭后统一隐藏，否则已
  装备格每帧被置回 Visible 悬浮在游戏画面上）；ui_system 移除空转的
  icons 隐藏循环（其 Without<InspectWidget> 查询恒空——图标实体同时挂
  两标记，审查确认空查询反使 disjoint 恒成立但意图落空）
- MINOR 标签框心：名字 (145,22)/行会 (145,48) + Anchor::CENTER + 8F
  （C# :2317-2324/:2343-2350）；名字文本=纯 Name、行会无 rank 只显示名
  （偏差附 #2607）；角色页性别换帧 340/341（C# :2474-2476，InspectPage
  标记 + icon_system 换帧）；ClassImage/纸娃娃/五按钮列入有意偏差
- NIT：死绑定清理、重复注释清理

门禁：Client-Bevy 299 全绿；ServerRust check+fmt 全绿；e2e run28
（修后断言区域随标签新位）背景 21.7/名字 213/武器格 22 全过。

+++ ServerRust/src/actors/world/mod.rs
@@ -11025,8 +11025,9 @@ fn send_inspect_packet(
         body.extend_from_slice(&eq.unique_id.to_le_bytes());
         body.extend_from_slice(&eq.item_index.to_le_bytes());
         // #2607：图标索引（ItemInfo.Image，Items 库位）——客户端无 item_index→image
-        // 信息表，服务端带上才能渲染 14 格网格图标（C# 客户端本地有 ItemInfos）
-        let image = eq.info.as_ref().map(|i| i.image).unwrap_or(0);
+        // 信息表，服务端带上才能渲染 14 格网格图标（C# 客户端本地有 ItemInfos）。
+        // ⚠ ItemInfo.image 是 u16——必须显式 as i32（客户端按 i32 读，审查 BLOCKER）
+        let image = eq.info.as_ref().map(|i| i.image).unwrap_or(0) as i32;
         body.extend_from_slice(&image.to_le_bytes());
         body.extend_from_slice(&(eq.current_dura as i32).to_le_bytes());
         body.extend_from_slice(&(eq.max_dura as i32).to_le_bytes());
